+++ examples/libadwaita/leaflet_sidebar.rs
@@ -58,6 +58,12 @@ impl SimpleComponent for App {
                     }
                 },
 
+                append = &gtk::Separator {
+                    set_orientation: gtk::Orientation::Vertical,
+                } -> {
+                    set_navigatable: false,
+                },
+
                 gtk::Box {
                     set_orientation: gtk::Orientation::Vertical,
                     set_hexpand: true,